+++ demo/coder.R
@@ -1,0 +1,9 @@
+#' Learn from echarty examples
+#' Additional information is available by clicking the "Info" button inside.
+#' From RStudio run 'demo(coder)'  after loading package 'echarty'
+
+devAskNewPage(ask = FALSE)
+if (interactive()) {
+  library(shiny)
+  shiny::runGist('https://gist.github.com/helgasoft/02b257429e78e138f87aefce14f7aebc')
+}